Extract view model and create preview
